requirements.txt
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1   
bleach==5.0.1
bokeh==3.0.2
certifi==2022.9.24       
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1        
contourpy==1.0.6
cv==1.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
flask-ngrok==0.0.25
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.0.0
keyring==23.11.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplot==0.1.9
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==9.0.0
MouseInfo==0.1.3
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.4
opencv-contrib-python==4.6.0.66
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
pkginfo==1.8.3
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.13.0
pyloco==0.0.139
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
readme-renderer==37.3
regex==2022.10.31
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
SimpleWebSocketServer==0.1.2
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data==0.6.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
twine==4.0.1
typing==3.7.4.3
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wikipedia==1.4.0
xyzservices==2022.9.0
yarg==0.1.9
zipp==3.10.0